--- README.md
@@ -1,17 +1,0 @@
-# React + Vite
-
-This template provides a minimal setup to get React working in Vite with HMR and some ESLint rules.
-
-Currently, two official plugins are available:
-
-- [@vitejs/plugin-react](https://github.com/vitejs/vite-plugin-react/blob/main/packages/plugin-react/README.md) uses [Babel](https://babeljs.io/) for Fast Refresh
-- [@vitejs/plugin-react-swc](https://github.com/vitejs/vite-plugin-react-swc) uses [SWC](https://swc.rs/) for Fast Refresh
-
-
-To start a project, you need to register an "npm run dev"
-Node Version: 10.4.0
-To install dependencies, the command in the terminal is "npm install" or "npm i" and npm install @emotion/react @emotion/styled
-
-
-
-then "npm dev" is enough
